Create string_manipulation.py
Instructions
Let's filter out the letter 'A' from our string.

Do the following for each character in the phrase.
If char is an 'A' or char is an 'a', print 'X', instead of char. Make sure to include the trailing comma.
Otherwise (else:), please print char, with the trailing comma.