Create comparisson.py
Multiple lists
It's also common to need to iterate over two lists at once. This is where the built-in zip function comes in handy.

zip will create pairs of elements when passed two lists, and will stop at the end of the shorter list.

zip can handle three or more lists as well!

Instructions
Compare each pair of elements and print the larger of the two.

?
Hint
a is an element from list_a and b is an element of list_b. Use an if statement to compare the two and print whichever is larger.